requirement.txt
cffi
cryptography
Django
djangorestframework
django-guardian
idna
mysqlclient
pyasn1
pycparser
pyOpenSSL
pytz
requests
rsa
six
whitenoise
